requirements.txt
streamlit
pandas
plotly
easyocr
Pillow
numpy
opencv-python-headless
langchain
langgraph
langchain-ollama
sqlite3-binary
